Separated Bancor v2.1 and v3
Separated v2.1 and v3 adapters into index.js and v3.js, respectively, and added hallmarks.
